backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
python-dotenv==1.0.0
langchain==0.0.267
openai==0.28.1
python-jose[cryptography]==3.3.0
